+++ components/Layout/index.tsx
@@ -1,4 +1,4 @@
-import Navbar from './Navbar';
+import Navbar from '../Navbar';
 
 const Layout = ({ children }) => (
   <div>
