+++ styles/index.scss
@@ -1,4 +1,3 @@
-@import "../components/Navbar.scss";
 html,
 body {
   padding: 0;
